Added a method for determining to add or not add closing tags
Per an email I recieved, someone asked if there is a method to output a CTML node without a closing tag (for elements such as `img`). Now there is! Just use the `AddClosingTag` method with a boolean argument, true to add one, false to remove it. Every node defaults to true with this change.