add `html` as Node to customize attributes
The `html` tag does accept global attributes. This was not possible
as the `html` tag was hard coded.

This change introduces the `html` tag as a dedicated Node and adds the
`head` and `body` nodes as childs.
In addition, `Node`s can not be queried for a named child returning the
first of such found.

Fixes issue #16.