Started rewrite/clean-up
CTML was written a good few years ago when I was pretty much first learning C++. I've changed my code style a lot, and also learned some new tricks about C++. In the interest of bettering the library code for contribution, as well as for future features, I've decided to clean up a bit. This is the start.