Removed uses of string view and reverted to C++11
I had a realization that C++17 is very new and thus is probably not as widely
supported as C++11. Thus I got rid of string view and reverted back to const
strings. I also had to change the options struct to use a constructor with all
default values rather than an initializer list as C++11 doesn't like it. Oops.