Added a parent node member to Node
Removed the selector remove child method in favor of an in place Remove method.
I am getting ready to add basic selector support for finding nodes within the
children of a node, thus removing by selector would be redundant when you can
just find the node and remove it in place. This also required a change where
AppendChild now has two definitions for taking in a reference as well as a
constant Node definition so that parents could be set accordingly.