Separate selector/Emmet abbreviation parsing
To try and make the Node class more lean and add more functionality regarding
selectors, I made a rudamentary selector parser in the function
`CTML::parse_selector` that should be more flexible and clearer than the
previous `_ParseClassesAndIDs` method in the Node class. I also added a flag
when compiling the test project for debug symbols to be generated for use in GDB
to ease finding issues.